% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdist2.R
\name{rdist}
\alias{rdist}
\title{Simulation d'une loi discrete'}
\usage{
rdist(x, p)
}
\arguments{
\item{x}{Numeric vector representing  the Values}

\item{p}{numeric vector representing probabies}

\item{n}{entier representing numbre of iteration}
}
\description{
Simulation d'une loi discrete'
}
